implementing review feedback, small refactor
