+++ backend/app/views/backoffice/project_developers/_form.html.erb
@@ -1,6 +1,5 @@
 <%= simple_form_for [:backoffice, project_developer] do |f| %>
   <%= f.error_notification %>
-  <%= f.error_notification message: f.object.errors[:base].to_sentence if f.object.errors[:base].present? %>
 
   <h2 class="mb-4 mt-6 text-gray-600 font-semibold text-xl">
     <%= t(".general") %>
@@ -17,7 +16,7 @@
   </div>
   <%= f.input :entity_legal_registration_number %>
 
-  <%= render "account_fields", f: f %>
+  <%= render "backoffice/accounts/common_fields", f: f %>
 
   <h2 class="mb-4 mt-6 text-gray-600 font-semibold text-xl">
     <%= t(".about_the_work") %>
