+++ backend/app/views/backoffice/project_developers/_section_status.html.erb
@@ -1,6 +1,5 @@
 <%= simple_form_for [:backoffice, @project_developer] do |f| %>
   <%= f.error_notification %>
-  <%= f.error_notification message: f.object.errors[:base].to_sentence if f.object.errors[:base].present? %>
 
   <h2 class="mb-2 text-gray-900 font-semibold">
     <%= t("backoffice.account.approval_status") %>
